requirements.txt
django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
ipython==8.23.0
drf-yasg==1.21.7
python-dotenv==1.0.1